Break out UI element into variables
Now, this here is an example of how to make the code easier to read.
The build function was very busy and has a lot of information in it
which makes it really hard to understand what it's trying to do. So what
I have done here is break out each UI element out of the main build
function and put it else where so there is a definite separation of
concerns. The main `build` function is now mainly concerned with
putting all the different elements together in order, as the glue, while
the what and how of constructing each individual element is the
responsiblity of some other component else where.
There are a lot of many different ways to accomplish this but the theme
is the "separation of responsibility". To make your code easily to
follow along, and testable, you want to make sure that the functionality
are divvy up into into units where each unit is only responsible for
just one thing. I'm happy to go through more examples of ways to do this
with you too.